# Support for cel.@block during policy composition