@TristonianJones 2024-10-29
@seirl FYI -- once cel.@block and this PR are landed, the import of cel-go will fix the recursion depth challenges you encountered with large policies.